+++ src/test/java/nextstep/subway/line/application/LineServiceTest.java
@@ -1,6 +1,5 @@
-package nextstep.subway.line.unit;
+package nextstep.subway.line.application;
 
-import nextstep.subway.line.application.LineService;
 import nextstep.subway.line.application.dto.SectionRequest;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
